src/Execution/Expression.test.ts
import "jest-extended";

import { Value } from "../AST";
import { executeExpr } from "./Expression";
import { ProgramScope, declare } from "./Scope";
import { DynamicTypeError } from "./TypeAssertions";

test("conditional expression test", () => {
  const scope: ProgramScope = [new Map()];

  expect(executeExpr(scope, {
    tag: "conditional",
    condition: { tag: "bool", value: true },
    trueExpr: { tag: "num", value: 2 },
    falseExpr: { tag: "num", value: 3 }
  })).toEqual(2);

  expect(scope).toEqual([new Map()]);

  expect(executeExpr(scope, {
    tag: "conditional",
    condition: { tag: "bool", value: false },
    trueExpr: { tag: "num", value: 2 },
    falseExpr: { tag: "num", value: 3 }
  })).toEqual(3);

  expect(scope).toEqual([new Map()]);

  expect(() => executeExpr(scope, {
    tag: "conditional",
    condition: { tag: "num", value: 0 },
    trueExpr: { tag: "num", value: 2 },
    falseExpr: { tag: "num", value: 3 }
  })).toThrowError(DynamicTypeError);

  expect(scope).toEqual([new Map()]);

  expect(executeExpr(scope, {
    tag: "conditional",
    condition: {
      tag: "and",
      leftSubexpr: { tag: "bool", value: true },
      rightSubexpr: { tag: "bool", value: false }
    },
    trueExpr: { tag: "num", value: 10 },
    falseExpr: { tag: "num", value: 100 }
  })).toEqual(100);

  expect(scope).toEqual([new Map()]);

  expect(executeExpr(scope, {
    tag: "conditional",
    condition: {
      tag: "or",
      leftSubexpr: {
        tag: "and",
        leftSubexpr: { tag: "bool", value: true },
        rightSubexpr: { tag: "bool", value: false }
      },
      rightSubexpr: {
        tag: "or",
        leftSubexpr: { tag: "bool", value: false },
        rightSubexpr: { tag: "bool", value: true }
      }
    },
    trueExpr: { tag: "num", value: 10 },
    falseExpr: { tag: "num", value: 100 }
  })).toEqual(10);

  expect(scope).toEqual([new Map()]);

  expect(() => executeExpr(scope, {
    tag: "conditional",
    condition: {
      tag: "or",
      leftSubexpr: {
        tag: "and",
        leftSubexpr: { tag: "bool", value: true },
        rightSubexpr: { tag: "bool", value: false }
      },
      rightSubexpr: {
        tag: "or",
        leftSubexpr: { tag: "num", value: 1 },
        rightSubexpr: { tag: "bool", value: true }
      }
    },
    trueExpr: { tag: "num", value: 10 },
    falseExpr: { tag: "num", value: 100 }
  })).toThrowError(DynamicTypeError);

  expect(scope).toEqual([new Map()]);

  expect(() => executeExpr(scope, {
    tag: "conditional",
    condition: {
      tag: "or",
      leftSubexpr: {
        tag: "and",
        leftSubexpr: { tag: "bool", value: true },
        rightSubexpr: { tag: "bool", value: false }
      },
      rightSubexpr: {
        tag: "or",
        leftSubexpr: { tag: "num", value: 1 },
        rightSubexpr: { tag: "bool", value: true }
      }
    },
    trueExpr: {
      tag: "plus",
      leftSubexpr: {
        tag: "times",
        leftSubexpr: { tag: "num", value: 3 },
        rightSubexpr: { tag: "num", value: 30 }
      },
      rightSubexpr: {
        tag: "exponent",
        leftSubexpr: { tag: "num", value: 1 },
        rightSubexpr: { tag: "bool", value: true }
      }
    },
    falseExpr: {
      tag: "minus",
      leftSubexpr: {
        tag: "divide",
        leftSubexpr: { tag: "num", value: 3 },
        rightSubexpr: { tag: "num", value: 30 }
      },
      rightSubexpr: {
        tag: "plus",
        leftSubexpr: { tag: "num", value: 1 },
        rightSubexpr: { tag: "num", value: 2 }
      }
    }
  })).toThrowError(DynamicTypeError);

  expect(scope).toEqual([new Map()]);

  expect(executeExpr(scope, {
    tag: "conditional",
    condition: {
      tag: "or",
      leftSubexpr: {
        tag: "and",
        leftSubexpr: { tag: "bool", value: true },
        rightSubexpr: { tag: "bool", value: false }
      },
      rightSubexpr: {
        tag: "or",
        leftSubexpr: { tag: "bool", value: false },
        rightSubexpr: { tag: "bool", value: true }
      }
    },
    trueExpr: {
      tag: "plus",
      leftSubexpr: {
        tag: "times",
        leftSubexpr: { tag: "num", value: 3 },
        rightSubexpr: { tag: "num", value: 30 }
      },
      rightSubexpr: {
        tag: "exponent",
        leftSubexpr: { tag: "num", value: 1 },
        rightSubexpr: { tag: "num", value: 2 }
      }
    },
    falseExpr: {
      tag: "minus",
      leftSubexpr: {
        tag: "divide",
        leftSubexpr: { tag: "num", value: 3 },
        rightSubexpr: { tag: "num", value: 30 }
      },
      rightSubexpr: {
        tag: "plus",
        leftSubexpr: { tag: "num", value: 1 },
        rightSubexpr: { tag: "num", value: 2 }
      }
    }
  })).toEqual(91)

  expect(scope).toEqual([new Map()]);

  declare("x", 1, scope);
  declare("y", true, scope);

  expect(() => executeExpr(scope, {
    tag: "conditional",
    condition: { tag: "var", name: "x" },
    trueExpr: { tag: "num", value: 2 },
    falseExpr: { tag: "num", value: 3 }
  })).toThrowError(DynamicTypeError);

  expect(scope).toEqual([new Map<string, Value>([["x", 1], ["y", true]])]);

  expect(executeExpr(scope, {
    tag: "conditional",
    condition: { tag: "var", name: "y" },
    trueExpr: { tag: "num", value: 2 },
    falseExpr: { tag: "num", value: 3 }
  })).toEqual(2);

  expect(scope).toEqual([new Map<string, Value>([["x", 1], ["y", true]])]);

  scope[0].delete("x");
  scope[0].delete("y");
});
